1.添加砍价系数

+++ sql/update.sql
@@ -410,7 +410,7 @@ ADD COLUMN `symbol` varchar(50) NOT NULL COMMENT '砍价口令' AFTER `close_tim
 /***帮砍日志*****/
 CREATE TABLE `plant_bargain_rate` (
   `product_id` int(11) NOT NULL COMMENT '用户id',
-  `own_rate` decimal(15, 2) NOT NULL COMMENT '本人砍价系数',
-  `new_rate` decimal(15, 2) NOT NULL COMMENT '新用户系数',
-  `old_rate` decimal(15, 2) NOT NULL COMMENT '老用户系数'
+  `own_rate` int(11) NOT NULL COMMENT '本人砍价系数',
+  `new_rate` int(11) NOT NULL COMMENT '新用户系数',
+  `old_rate` int(11) NOT NULL COMMENT '老用户系数'
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
